requirements.txt
python-docx>=1.1.0
# ============================================================================
# PDF转换器项目依赖文件
# ============================================================================

# 核心PDF处理库
pdf2docx>=0.5.8          # PDF转DOCX转换
pdfplumber>=0.9.0        # PDF文本提取和分析
reportlab>=4.0.0         # PDF生成和处理

# Office文档处理
python-pptx>=0.6.21      # PowerPoint文档操作

# 图像处理
Pillow>=10.0.0           # 图像处理和格式转换

# OCR文字识别已移除
# pdf2image>=1.16.3        # PDF转图像已移除（主要用于OCR功能）

# GUI框架（Python内置，无需安装）
# tkinter - 图形用户界面

# ============================================================================
# 系统依赖要求（需要手动安装）
# ============================================================================

# 1. Tesseract OCR引擎已移除

# 2. Poppler PDF工具集（用于PDF转图像）
#    Windows: https://github.com/oschwartz10612/poppler-windows/releases/
#    推荐版本: 24.08.0或更高
#    安装步骤:
#    - 下载并解压到 C:\poppler
#    - 将 C:\poppler\Library\bin 添加到系统PATH环境变量
#    - 重启应用程序使环境变量生效

# ============================================================================
# 安装说明
# ============================================================================

# 1. 安装Python依赖:
#    pip install -r requirements.txt

# 2. 或使用虚拟环境（推荐）:
#    python -m venv .venv
#    .venv\Scripts\activate  # Windows
#    pip install -r requirements.txt

# 3. 验证安装:
#    python -c "import pdf2docx, pdfplumber; print('依赖安装成功')"

# ============================================================================
# 故障排除
# ============================================================================

# OCR功能已移除
